# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","operationId":"get_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"No-op: Supabase session is managed client-side.","operationId":"logout_auth_logout_post","responses":{"204":{"description":"Successful Response"}}}},"/properties":{"get":{"tags":["properties"],"summary":"List Properties","operationId":"list_properties_properties_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PropertyResponse"},"type":"array","title":"Response List Properties Properties Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["properties"],"summary":"Create Property","operationId":"create_property_properties_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/properties/{property_id}":{"get":{"tags":["properties"],"summary":"Get Property","operationId":"get_property_properties__property_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["properties"],"summary":"Update Property","operationId":"update_property_properties__property_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["properties"],"summary":"Delete Property","operationId":"delete_property_properties__property_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/properties/{property_id}/rooms":{"post":{"tags":["properties"],"summary":"Create Room","operationId":"create_room_properties__property_id__rooms_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_room_properties__property_id__rooms_post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/properties/{property_id}/rooms/{room_id}":{"get":{"tags":["properties"],"summary":"Get Room","operationId":"get_room_properties__property_id__rooms__room_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}},{"name":"room_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Room Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["properties"],"summary":"Update Room","operationId":"update_room_properties__property_id__rooms__room_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}},{"name":"room_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Room Id"}},{"name":"label","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_update_room_properties__property_id__rooms__room_id__put"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["properties"],"summary":"Delete Room","operationId":"delete_room_properties__property_id__rooms__room_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}},{"name":"room_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Room Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/styles":{"get":{"tags":["styles"],"summary":"List Styles","operationId":"list_styles_styles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DesignStyle"},"type":"array","title":"Response List Styles Styles Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/generation":{"post":{"tags":["generation"],"summary":"Start Generation For Room","description":"Submit and generate an AI generation request for a room in the property.","operationId":"start_generation_for_room_generation_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GenerationJobCreate"},"type":"array","title":"Payload"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GenerationJobResponse"},"type":"array","title":"Response Start Generation For Room Generation Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/generation/property/{property_id}":{"get":{"tags":["generation"],"summary":"List Jobs For Property","description":"See all generation jobs for a property.","operationId":"list_jobs_for_property_generation_property__property_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Property Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenerationJobResponse"},"title":"Response List Jobs For Property Generation Property  Property Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generation/jobs/{job_id}":{"get":{"tags":["generation"],"summary":"Get Job","description":"Get the specific genertaion job, if it exists","operationId":"get_job_generation_jobs__job_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationJobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}}},"components":{"schemas":{"Body_create_room_properties__property_id__rooms_post":{"properties":{"image":{"type":"string","contentMediaType":"application/octet-stream","title":"Image"},"label":{"type":"string","title":"Label"}},"type":"object","required":["image","label"],"title":"Body_create_room_properties__property_id__rooms_post"},"Body_update_room_properties__property_id__rooms__room_id__put":{"properties":{"image":{"type":"string","contentMediaType":"application/octet-stream","title":"Image"}},"type":"object","title":"Body_update_room_properties__property_id__rooms__room_id__put"},"DesignStyle":{"properties":{"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"},"description":{"type":"string","title":"Description"},"image_urls":{"additionalProperties":{"type":"string"},"type":"object","title":"Image Urls"}},"type":"object","required":["name","slug","description","image_urls"],"title":"DesignStyle"},"GenerationJobCreate":{"properties":{"room_id":{"type":"string","format":"uuid","title":"Room Id"},"original_job_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Original Job Id"},"style":{"type":"string","title":"Style"},"extra_context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extra Context"}},"type":"object","required":["room_id","style"],"title":"GenerationJobCreate"},"GenerationJobResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"room_id":{"type":"string","format":"uuid","title":"Room Id"},"style":{"type":"string","title":"Style"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","room_id","style","image_url","error_message","completed_at","created_at","updated_at"],"title":"GenerationJobResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","const":"ok","title":"Status","default":"ok"}},"type":"object","title":"HealthCheckResponse"},"PropertyCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Address"}},"type":"object","required":["name"],"title":"PropertyCreate"},"PropertyResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"owner_id":{"type":"string","format":"uuid","title":"Owner Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"rooms":{"items":{"$ref":"#/components/schemas/RoomResponse"},"type":"array","title":"Rooms"}},"type":"object","required":["id","name","description","address","owner_id","created_at","updated_at","rooms"],"title":"PropertyResponse"},"PropertyUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Address"}},"type":"object","title":"PropertyUpdate"},"RoomResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"property_id":{"type":"string","format":"uuid","title":"Property Id"},"label":{"type":"string","title":"Label"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"generation_jobs":{"items":{"$ref":"#/components/schemas/GenerationJobResponse"},"type":"array","title":"Generation Jobs"}},"type":"object","required":["id","property_id","label","image_url","created_at","updated_at","generation_jobs"],"title":"RoomResponse","description":"Schema used in the API Response"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"picture":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Picture"},"balance":{"type":"number","title":"Balance"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","name","picture","balance","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}